Verify Microprofile, fallback doesn't work on Native mode ++ Refactor
The test to verify "https://issues.redhat.com/browse/QUARKUS-697:Microprofile, fallback doesn't work on Native mode" was disabled and these changes are to enable them.
I've also refactored the module to rename the packages and run the OCP test on baremetal too.
