+++ microprofile/src/main/resources/application.properties
@@ -4,4 +4,5 @@ quarkus.jaeger.service-name=test-traced-service
 quarkus.jaeger.sampler-type=const
 quarkus.jaeger.sampler-param=1
 quarkus.jaeger.endpoint=http://localhost:14268/api/traces
-io.quarkus.ts.openshift.microprofile.HelloClient/mp-rest/url=http://localhost:${quarkus.http.port}
+
+io.quarkus.ts.microprofile.HelloClient/mp-rest/url=http://localhost:${quarkus.http.port}
